+++ core/strategies/__init__.py
@@ -7,9 +7,15 @@ from .dga import DGA
 from .fedlabels import FedLabels
 
 def select_strategy(strategy):
+    ''' Selects the aggregation strategy class
+    
+    NOTE: FedProx uses FedAvg weights during aggregation, 
+    which are proportional to the number of samples in 
+    each client.
+    '''
     if strategy.lower() == 'dga':
         return DGA
-    elif strategy.lower() == 'fedavg':
+    elif strategy.lower() in ['fedavg', 'fedprox']:
         return FedAvg
     elif strategy.lower() == 'fedlabels':
         return FedLabels